Move library include files to the include/upa directory.
