+++ src/url_percent_encode.cpp
@@ -7,7 +7,7 @@
 // Copyright 2013 The Chromium Authors. All rights reserved.
 //
 
-#include "url_percent_encode.h"
+#include "upa/url_percent_encode.h"
 
 namespace upa { // NOLINT(modernize-concat-nested-namespaces)
