+++ test/test-ipv4.cpp
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 //
 
-#include "url.h"
+#include "upa/url.h"
 #include "doctest-main.h"
