+++ test/test-str_arg.cpp
@@ -7,7 +7,7 @@
 // NOTE: To pass test this file must just compile without errors.
 //
 
-#include "str_arg.h"
+#include "upa/str_arg.h"
 //#include "doctest-main.h"
