+++ test/test-url_host.cpp
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 //
 
-#include "url_host.h"
+#include "upa/url_host.h"
 #include "doctest-main.h"
