+++ test/test-url_percent_encode.cpp
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 //
 
-#include "url_percent_encode.h"
+#include "upa/url_percent_encode.h"
 #include "doctest-main.h"
 #include "test-utils.h"
