+++ test/test-util.cpp
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 //
 
-#include "util.h"
+#include "upa/util.h"
 #include "doctest-main.h"
